Created required functions (#115)
* Created required functions

* exported needed function

* ran prettier

* Update tasks/utils/github.js


* Update tasks/utils/github.js


---------
